Refactor: Extract Ical-related methods to a separate class
