+++ java/code/src/com/suse/manager/maintenance/MaintenanceManager.java
@@ -17,8 +17,6 @@ package com.suse.manager.maintenance;
 import static com.redhat.rhn.domain.role.RoleFactory.ORG_ADMIN;
 import static java.util.Collections.emptySet;
 import static java.util.Optional.empty;
-import static java.util.Optional.of;
-import static java.util.Optional.ofNullable;
 import static java.util.stream.Collectors.joining;
 import static java.util.stream.Collectors.toList;
 import static java.util.stream.Collectors.toSet;
@@ -49,7 +47,6 @@ import com.suse.manager.reactor.messaging.ApplyStatesEventMessage;
 import com.suse.manager.utils.HttpHelper;
 import com.suse.utils.Opt;
 
-import org.apache.commons.lang3.tuple.Pair;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.http.HttpResponse;
 import org.apache.http.HttpStatus;
@@ -58,10 +55,6 @@ import org.apache.http.StatusLine;
 import org.apache.log4j.Logger;
 
 import java.io.IOException;
-import java.io.Reader;
-import java.io.StringReader;
-import java.time.Duration;
-import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
@@ -74,19 +67,14 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
-import net.fortuna.ical4j.data.CalendarBuilder;
-import net.fortuna.ical4j.data.ParserException;
 import net.fortuna.ical4j.filter.Filter;
 import net.fortuna.ical4j.filter.HasPropertyRule;
 import net.fortuna.ical4j.filter.PeriodRule;
 import net.fortuna.ical4j.model.Calendar;
 import net.fortuna.ical4j.model.Component;
-import net.fortuna.ical4j.model.ComponentList;
 import net.fortuna.ical4j.model.DateTime;
 import net.fortuna.ical4j.model.Period;
-import net.fortuna.ical4j.model.PeriodList;
 import net.fortuna.ical4j.model.component.CalendarComponent;
 import net.fortuna.ical4j.model.property.Summary;
 
@@ -100,6 +88,7 @@ public class MaintenanceManager {
 
     private ScheduleFactory scheduleFactory;
     private CalendarFactory calendarFactory;
+    private IcalUtils icalUtils;
 
     /**
      * Constructor.
@@ -107,6 +96,7 @@ public class MaintenanceManager {
     public MaintenanceManager() {
         scheduleFactory = new ScheduleFactory();
         calendarFactory = new CalendarFactory();
+        icalUtils = new IcalUtils();
     }
 
     /**
@@ -161,89 +151,7 @@ public class MaintenanceManager {
         }
 
         MaintenanceSchedule schedule = schedules.iterator().next();
-        return calculateMaintenanceWindows(schedule);
-    }
-
-    /**
-     * Given MaintenanceSchedule calculate upcoming maintenance windows
-     *
-     * The windows are returned as a list of triples consisting of:
-     * - window start date as a human-readable string
-     * - window end date as a human-readable string
-     * - start date as number of milliseconds since the epoch
-     *
-     * The formatting is done by {@link LocalizationService}.
-     *
-     * The upper limit of returned maintenance windows is currently hardcoded to 10.
-     *
-     * @param schedule the given MaintenanceSchedule
-     * @return the optional upcoming maintenance windows
-     */
-    public Optional<List<MaintenanceWindowData>> calculateMaintenanceWindows(MaintenanceSchedule schedule) {
-        Optional<String> multiScheduleName = getScheduleNameForMulti(schedule);
-
-        Stream<Pair<Instant, Instant>> periodStream = schedule.getCalendarOpt()
-                .flatMap(c -> parseCalendar(c))
-                .map(c -> calculateUpcomingPeriods(c, multiScheduleName, Instant.now(), 10))
-                .orElseGet(Stream::empty);
-
-        List<MaintenanceWindowData> result = periodStream
-                .map(p -> new MaintenanceWindowData(p.getLeft(), p.getRight()))
-                .collect(toList());
-        return of(result);
-    }
-
-    /**
-     * Convenience method: return schedule name if the schedule type is MULTI, return empty otherwise
-     * @param schedule the schedule
-     * @return optional of schedule name
-     */
-    private static Optional<String> getScheduleNameForMulti(MaintenanceSchedule schedule) {
-        if (schedule.getScheduleType() == ScheduleType.MULTI) {
-            return of(schedule.getName());
-        }
-        return empty();
-    }
-
-    /**
-     * Calculate upcoming maintenance windows starting from given date based on calendar and optional filter name
-     * (in case we're dealing with MULTI calendar and want to filter only events we're interested in).
-     *
-     * The algorithm only checks maintenance windows within roughly a year and a month since the startDate.
-     *
-     * @param calendar the {@link Calendar}
-     * @param eventName for MULTI calendars: only deal with events with this name, filter out the rest
-     * @param startDate the start date
-     * @param limit upper limit of maintenance windows to return
-     * @return the list of upcoming maintenance windows
-     */
-    public Stream<Pair<Instant, Instant>> calculateUpcomingPeriods(Calendar calendar, Optional<String> eventName,
-            Instant startDate, int limit) {
-        ComponentList<CalendarComponent> allEvents = calendar.getComponents(Component.VEVENT);
-
-        Collection<CalendarComponent> filteredEvents = eventName.map(name -> {
-            Predicate<CalendarComponent> summary = c -> c.getProperty("SUMMARY").equals(name);
-            Predicate<CalendarComponent>[] ps = new Predicate[]{summary};
-            Filter<CalendarComponent> filter = new Filter<>(ps, Filter.MATCH_ALL);
-            return filter.filter(allEvents);
-        }).orElse(allEvents);
-
-        // we will look a year and month to the future
-        Period period = new Period(new DateTime(startDate.toEpochMilli()), Duration.ofDays(365 + 31));
-
-        List<PeriodList> periodLists = filteredEvents.stream()
-                .map(c -> c.calculateRecurrenceSet(period))
-                .filter(l -> !l.isEmpty())
-                .collect(toList());
-
-        Stream<Pair<Instant, Instant>> sortedLimited = periodLists.stream()
-                .map(pl -> pl.stream())
-                .reduce(Stream.empty(), Stream::concat)
-                .sorted()
-                .limit(limit)
-                .map(p -> Pair.of(p.getStart().toInstant(), p.getRangeEnd().toInstant()));
-
-        return sortedLimited;
+        return icalUtils.calculateUpcomingMaintenanceWindows(schedule);
     }
 
     /**
@@ -290,7 +198,7 @@ public class MaintenanceManager {
         return listSchedulesOfSystems(systemIds).stream()
                 .filter(schedule -> {
                     Collection<CalendarComponent> events = getScheduleEventsAtDate(date, schedule, schedule
-                            .getCalendarOpt().flatMap(c -> parseCalendar(c)));
+                            .getCalendarOpt().flatMap(c -> icalUtils.parseCalendar(c)));
                     return events.isEmpty();
                 })
                 .collect(toSet());
@@ -620,7 +528,7 @@ public class MaintenanceManager {
         }
         List<Server> servers = ServerFactory.lookupByIdsAndOrg(withMaintenanceActions, user.getOrg());
 
-        Optional<Calendar> calendarOpt = schedule.getCalendarOpt().flatMap(c -> parseCalendar(c));
+        Optional<Calendar> calendarOpt = schedule.getCalendarOpt().flatMap(c -> icalUtils.parseCalendar(c));
 
         List<ActionStatus> pending = new LinkedList<>();
         pending.add(ActionFactory.STATUS_PICKED_UP);
@@ -671,30 +579,6 @@ public class MaintenanceManager {
         return new RescheduleResult(schedule.getName(), false);
     }
 
-    private Optional<Calendar> parseCalendar(MaintenanceCalendar calendarIn) {
-        return parseCalendar(new StringReader(calendarIn.getIcal()));
-    }
-
-    /**
-     * Read calendar using given reader and parse it
-     *
-     * Public for testing.
-     *
-     * @param calendarReader the reader
-     * @return the parsed calendar or empty, if there was a problem parsing the calendar
-     */
-    public Optional<Calendar> parseCalendar(Reader calendarReader) {
-        CalendarBuilder builder = new CalendarBuilder();
-        Calendar calendar = null;
-        try {
-            calendar = builder.build(calendarReader);
-        }
-        catch (IOException | ParserException e) {
-            log.error("Unable to build the calendar from reader: " + calendarReader, e);
-        }
-        return ofNullable(calendar);
-    }
-
     /**
      * Check if provided action is inside of a maintenance window
      *
@@ -722,7 +606,7 @@ public class MaintenanceManager {
 
     private Collection<CalendarComponent> getCalendarForNow(MaintenanceSchedule ms) {
         return ms.getCalendarOpt()
-                .map(cal -> getScheduleEventsAtDate(new Date(), ms, parseCalendar(cal)))
+                .map(cal -> getScheduleEventsAtDate(new Date(), ms, icalUtils.parseCalendar(cal)))
                 .orElse(Collections.emptyList());
     }
